requirements.txt
streamlit
pysqlite3
Pillow
pandas
streamlit-authenticator
